Update - eks cluster

+++ eks/destroy.sh
@@ -21,16 +21,20 @@ OIDC_PROVIDER=$(aws iam list-open-id-connect-providers --query OpenIDConnectProv
 #### Deleting individuals CF stacks
 if  grep STACK_NAME vars.sh >/dev/null 2>&1
 then
-aws cloudformation delete-stack --stack-name $WORKERNODE_STACK_NAME
-aws cloudformation wait stack-delete-complete --stack-name  $WORKERNODE_STACK_NAME
 aws cloudformation delete-stack --stack-name $CLUSTER_STACK_NAME
 aws cloudformation wait stack-delete-complete --stack-name $CLUSTER_STACK_NAME
 echo -e '\n \t' "EKS Stack has been removed"
 else
    echo -e '\n \t' "No Stack Name Found in vars.sh file"
 fi
+
+#### Delete EBS CSI role & policy
+if aws iam get-role --role-name ${CLUSTER_NAME}_EKS_EBS_role --output text >/dev/null 2>&1
+then
 aws iam detach-role-policy --policy-arn arn:aws:iam::$ACCID:policy/${CLUSTER_NAME}_EKS_EBS_Policy --role-name ${CLUSTER_NAME}_EKS_EBS_role
 aws iam delete-policy --policy-arn arn:aws:iam::$ACCID:policy/${CLUSTER_NAME}_EKS_EBS_Policy
 aws iam delete-open-id-connect-provider --open-id-connect-provider-arn $OIDC_PROVIDER
 aws iam delete-role --role-name ${CLUSTER_NAME}_EKS_EBS_role
+fi
+
 rm vars.sh
